Add structure to html
The index.html only had a container before. While
the grid should be created using dom manipulation,
that is only because it will be dynamic and would
be a hassle to make 10x10 boxes for a starting
grid.

Realizing this, the index.html should have the
rest of the components since they will be static
and never actually change (ie Buttons and other
containers).

The createGrid function has also been updated
since the grid has it's own container and no
longer needs an instance of the body's container.